fix prop command
